Modify(leesuhyung/spec): 스펙 수정중 사진이 수정되지 않는 경우와 기본 이미지로 변경했을 때 구분 #99

+++ mohae/src/specs/specs.controller.ts
@@ -175,8 +175,9 @@ export class SpecsController {
       updateSpecdto,
       specPhotoUrls,
     );
-
-    await this.awsService.deleteSpecS3Object(originSpecPhotoUrls);
+    if (originSpecPhotoUrls) {
+      await this.awsService.deleteSpecS3Object(originSpecPhotoUrls);
+    }
 
     return {
       msg: '성공적으로 스팩이 수정되었습니다.',
